rockchip: add WXY OEC / OECT support

+++ scripts/gen_image_generic.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 # Copyright (C) 2006-2012 OpenWrt.org
 set -e -x
-if [ $# -ne 5 ] && [ $# -ne 6 ]; then
-    echo "SYNTAX: $0 <file> <kernel size> <kernel directory> <rootfs size> <rootfs image> [<align>]"
+if [ $# -lt 5 ] || [ $# -gt 7 ]; then
+    echo "SYNTAX: $0 <file> <kernel size> <kernel directory> <rootfs size> <rootfs image> [<align>] [<logical-sector-size>]"
     exit 1
 fi
 
@@ -12,47 +12,58 @@ KERNELDIR="$3"
 ROOTFSSIZE="$4"
 ROOTFSIMAGE="$5"
 ALIGN="$6"
+LOGICAL_SECTOR_SIZE="${7:-512}"
+
+if [ "$LOGICAL_SECTOR_SIZE" -lt 512 ] || [ $((LOGICAL_SECTOR_SIZE % 512)) -ne 0 ]; then
+    echo "Invalid logical sector size: $LOGICAL_SECTOR_SIZE"
+    exit 1
+fi
 
 rm -f "$OUTPUT"
 
 head=16
 sect=63
+KERNELPART="${KERNELSIZE}m"
+[ -n "$KERNEL_OFFSET" ] && KERNELPART="${KERNELPART}@${KERNEL_OFFSET}k"
 
 # create partition table
-set $(ptgen -o "$OUTPUT" -h $head -s $sect ${GUID:+-g} -p "${KERNELSIZE}m" -p "${ROOTFSSIZE}m" ${ALIGN:+-l $ALIGN} ${SIGNATURE:+-S 0x$SIGNATURE} ${GUID:+-G $GUID})
+if [ -n "$GUID" ] && [ -n "$PURE_GPT" ]; then
+    set $(ptgen -o "$OUTPUT" -h $head -s $sect -g -D -p "$KERNELPART" -p "${ROOTFSSIZE}m" ${ALIGN:+-l $ALIGN} -G "$GUID")
+    GPT_DISK_SIZE=$((($3 + $4 + 1023) / 1024 + ${GPT_PADDING_KB:-1024}))
+    set $(ptgen -o "$OUTPUT" -h $head -s $sect -g -D -d "$GPT_DISK_SIZE" -p "$KERNELPART" -p "${ROOTFSSIZE}m" ${ALIGN:+-l $ALIGN} -G "$GUID")
+else
+    set $(ptgen -o "$OUTPUT" -h $head -s $sect ${GUID:+-g} -p "$KERNELPART" -p "${ROOTFSSIZE}m" ${ALIGN:+-l $ALIGN} ${SIGNATURE:+-S 0x$SIGNATURE} ${GUID:+-G $GUID})
+fi
 
 KERNELOFFSET="$(($1 / 512))"
 KERNELSIZE="$2"
 ROOTFSOFFSET="$(($3 / 512))"
 ROOTFSSIZE="$(($4 / 512))"
-
-# Using mcopy -s ... is using READDIR(3) to iterate through the directory
-# entries, hence they end up in the FAT filesystem in traversal order which
-# breaks reproducibility.
-# Implement recursive copy with reproducible order.
-dos_dircopy() {
-  local entry
-  local baseentry
-  for entry in "$1"/* ; do
-    if [ -f "$entry" ]; then
-      mcopy -i "$OUTPUT.kernel" "$entry" ::"$2"
-    elif [ -d "$entry" ]; then
-      baseentry="$(basename "$entry")"
-      mmd -i "$OUTPUT.kernel" ::"$2""$baseentry"
-      dos_dircopy "$entry" "$2""$baseentry"/
-    fi
-  done
-}
+ROOTFSOFFSET_ALT="$((ROOTFSOFFSET * (LOGICAL_SECTOR_SIZE / 512)))"
 
 [ -n "$PADDING" ] && dd if=/dev/zero of="$OUTPUT" bs=512 seek="$ROOTFSOFFSET" conv=notrunc count="$ROOTFSSIZE"
 dd if="$ROOTFSIMAGE" of="$OUTPUT" bs=512 seek="$ROOTFSOFFSET" conv=notrunc
 
-if [ -n "$GUID" ]; then
-    [ -n "$PADDING" ] && dd if=/dev/zero of="$OUTPUT" bs=512 seek="$((ROOTFSOFFSET + ROOTFSSIZE))" conv=notrunc count="$sect"
-    mkfs.fat --invariant -n kernel -C "$OUTPUT.kernel" -S 512 "$((KERNELSIZE / 1024))"
-    LC_ALL=C dos_dircopy "$KERNELDIR" /
-else
-    make_ext4fs -J -L kernel -l "$KERNELSIZE" ${SOURCE_DATE_EPOCH:+-T ${SOURCE_DATE_EPOCH}} "$OUTPUT.kernel" "$KERNELDIR"
+if [ "$ROOTFSOFFSET_ALT" -ne "$ROOTFSOFFSET" ]; then
+    [ -n "$PADDING" ] && dd if=/dev/zero of="$OUTPUT" bs=512 seek="$ROOTFSOFFSET_ALT" conv=notrunc count="$ROOTFSSIZE"
+    dd if="$ROOTFSIMAGE" of="$OUTPUT" bs=512 seek="$ROOTFSOFFSET_ALT" conv=notrunc
 fi
-dd if="$OUTPUT.kernel" of="$OUTPUT" bs=512 seek="$KERNELOFFSET" conv=notrunc
-rm -f "$OUTPUT.kernel"
+
+if [ -n "$KERNEL_IMAGE" ]; then
+    KERNEL_IMAGE_SIZE=$(stat -c%s "$KERNEL_IMAGE")
+    if [ "$KERNEL_IMAGE_SIZE" -gt "$KERNELSIZE" ]; then
+        echo "Kernel image exceeds configured kernel partition size"
+        exit 1
+    fi
+    dd if="$KERNEL_IMAGE" of="$OUTPUT" bs=512 seek="$KERNELOFFSET" conv=notrunc
+elif [ -n "$GUID" ]; then
+    [ -n "$PADDING" ] && dd if=/dev/zero of="$OUTPUT" bs=512 seek="$((ROOTFSOFFSET + ROOTFSSIZE))" conv=notrunc count="$sect"
+    mkfs.fat -n kernel -C "$OUTPUT.kernel" -S 512 "$((KERNELSIZE / 1024))"
+    mcopy -s -i "$OUTPUT.kernel" "$KERNELDIR"/* ::/
+else
+    make_ext4fs -J -L kernel -l "$KERNELSIZE" "$OUTPUT.kernel" "$KERNELDIR"
+fi
+if [ -z "$KERNEL_IMAGE" ]; then
+    dd if="$OUTPUT.kernel" of="$OUTPUT" bs=512 seek="$KERNELOFFSET" conv=notrunc
+    rm -f "$OUTPUT.kernel"
+fi
